Fix compilation issues after the DS move
