+++ lyo-client/src/main/java/org/eclipse/lyo/client/RootServicesHelper.java
@@ -16,7 +16,7 @@
  *     Samuel Padgett      - add request consumer key and OAuth approval module URLs
  *     Samuel Padgett      - handle trailing '/' in baseUrl
  *******************************************************************************/
-package org.eclipse.lyo.client.jazz;
+package org.eclipse.lyo.client;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -27,8 +27,6 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriBuilder;
 
-import org.eclipse.lyo.client.OSLCConstants;
-import org.eclipse.lyo.client.OslcClient;
 import org.eclipse.lyo.client.exception.ResourceNotFoundException;
 import org.eclipse.lyo.client.exception.RootServicesException;
 import org.apache.http.HttpEntity;
